+++ util-stats/src/main/scala/com/twitter/finagle/stats/VerbosityAdjustingStatsReceiver.scala
@@ -9,16 +9,16 @@ class VerbosityAdjustingStatsReceiver(
   defaultVerbosity: Verbosity)
     extends StatsReceiverProxy {
 
-  override def stat(histogramSchema: HistogramSchema): Stat = {
-    self.stat(HistogramSchema(histogramSchema.metricBuilder.withVerbosity(defaultVerbosity)))
+  override def stat(metricBuilder: MetricBuilder): Stat = {
+    self.stat(metricBuilder.withVerbosity(defaultVerbosity))
   }
 
-  override def counter(counterSchema: CounterSchema): Counter = {
-    self.counter(CounterSchema(counterSchema.metricBuilder.withVerbosity(defaultVerbosity)))
+  override def counter(metricBuilder: MetricBuilder): Counter = {
+    self.counter(metricBuilder.withVerbosity(defaultVerbosity))
   }
 
-  override def addGauge(gaugeSchema: GaugeSchema)(f: => Float): Gauge = {
-    self.addGauge(GaugeSchema(gaugeSchema.metricBuilder.withVerbosity(defaultVerbosity)))(f)
+  override def addGauge(metricBuilder: MetricBuilder)(f: => Float): Gauge = {
+    self.addGauge(metricBuilder.withVerbosity(defaultVerbosity))(f)
   }
 
 }