+++ util-stats/src/test/scala/com/twitter/finagle/stats/InMemoryStatsReceiverTest.scala
@@ -1,5 +1,6 @@
 package com.twitter.finagle.stats
 
+import com.twitter.finagle.stats.MetricBuilder.{CounterType, GaugeType, HistogramType}
 import com.twitter.finagle.stats.exp.{Expression, ExpressionSchema}
 import java.io.{ByteArrayOutputStream, PrintStream}
 import java.nio.charset.StandardCharsets
@@ -124,7 +125,8 @@ class InMemoryStatsReceiverTest extends AnyFunSuite with Eventually with Integra
     val stats = new InMemoryStatsReceiver()
     ExpressionSchema(
       "a",
-      Expression(CounterSchema(MetricBuilder(name = Seq("counter"), statsReceiver = stats))))
+      Expression(
+        MetricBuilder(name = Seq("counter"), metricType = CounterType, statsReceiver = stats)))
     stats.expressions.contains("a")
   }
 
@@ -326,11 +328,11 @@ class InMemoryStatsReceiverTest extends AnyFunSuite with Eventually with Integra
   test("creating metrics stores their metadata and clear()ing should remove that metadata") {
     val stats = new InMemoryStatsReceiver()
     stats.addGauge("coolGauge") { 3 }
-    assert(stats.schemas(Seq("coolGauge")).isInstanceOf[MetricSchema])
+    assert(stats.schemas(Seq("coolGauge")).isInstanceOf[MetricBuilder])
     stats.counter("sweetCounter")
-    assert(stats.schemas(Seq("sweetCounter")).isInstanceOf[MetricSchema])
+    assert(stats.schemas(Seq("sweetCounter")).isInstanceOf[MetricBuilder])
     stats.stat("radHisto")
-    assert(stats.schemas(Seq("radHisto")).isInstanceOf[MetricSchema])
+    assert(stats.schemas(Seq("radHisto")).isInstanceOf[MetricBuilder])
 
     stats.clear()
     assert(stats.schemas.isEmpty)
@@ -353,13 +355,13 @@ class InMemoryStatsReceiverTest extends AnyFunSuite with Eventually with Integra
       assert(parts.length == 3)
       assert(
         parts(
-          0) == "coolGauge GaugeSchema(MetricBuilder(false, No description provided, Unspecified, NoRoleSpecified, Verbosity(default), None, coolGauge, List(), None, Vector(), InMemoryStatsReceiver))")
+          0) == "coolGauge MetricBuilder(false, No description provided, Unspecified, NoRoleSpecified, Verbosity(default), None, coolGauge, List(), None, Vector(), GaugeType, InMemoryStatsReceiver)")
       assert(
         parts(
-          1) == "rad/histo HistogramSchema(MetricBuilder(false, No description provided, Unspecified, NoRoleSpecified, Verbosity(default), None, rad/histo, List(), None, Vector(), InMemoryStatsReceiver/rad))")
+          1) == "rad/histo MetricBuilder(false, No description provided, Unspecified, NoRoleSpecified, Verbosity(default), None, rad/histo, List(), None, Vector(), HistogramType, InMemoryStatsReceiver/rad)")
       assert(
         parts(
-          2) == "sweet/counter CounterSchema(MetricBuilder(false, No description provided, Unspecified, NoRoleSpecified, Verbosity(default), None, sweet/counter, List(), None, Vector(), InMemoryStatsReceiver))")
+          2) == "sweet/counter MetricBuilder(false, No description provided, Unspecified, NoRoleSpecified, Verbosity(default), None, sweet/counter, List(), None, Vector(), CounterType, InMemoryStatsReceiver)")
     } finally {
       ps.close()
     }
@@ -368,9 +370,12 @@ class InMemoryStatsReceiverTest extends AnyFunSuite with Eventually with Integra
   test("expression are reloaded with fully scoped names") {
     val sr = new InMemoryStatsReceiver
 
-    val aSchema = CounterSchema(MetricBuilder(name = Seq("a"), statsReceiver = sr).withKernel)
-    val bSchema = HistogramSchema(MetricBuilder(name = Seq("b"), statsReceiver = sr).withKernel)
-    val cSchema = GaugeSchema(MetricBuilder(name = Seq("c"), statsReceiver = sr).withKernel)
+    val aSchema =
+      MetricBuilder(name = Seq("a"), metricType = CounterType, statsReceiver = sr).withKernel
+    val bSchema =
+      MetricBuilder(name = Seq("b"), metricType = HistogramType, statsReceiver = sr).withKernel
+    val cSchema =
+      MetricBuilder(name = Seq("c"), metricType = GaugeType, statsReceiver = sr).withKernel
 
     val expression = ExpressionSchema(
       "test_expression",
@@ -382,9 +387,12 @@ class InMemoryStatsReceiverTest extends AnyFunSuite with Eventually with Integra
     val cGauge = sr.scope(("test")).addGauge(cSchema) { 1 }
 
     // what we expected as hydrated metric builders
-    val aaSchema = CounterSchema(MetricBuilder(name = Seq("test", "a"), statsReceiver = sr))
-    val bbSchema = HistogramSchema(MetricBuilder(name = Seq("test", "b"), statsReceiver = sr))
-    val ccSchema = GaugeSchema(MetricBuilder(name = Seq("test", "c"), statsReceiver = sr))
+    val aaSchema =
+      MetricBuilder(name = Seq("test", "a"), metricType = CounterType, statsReceiver = sr)
+    val bbSchema =
+      MetricBuilder(name = Seq("test", "b"), metricType = HistogramType, statsReceiver = sr)
+    val ccSchema =
+      MetricBuilder(name = Seq("test", "c"), metricType = GaugeType, statsReceiver = sr)
 
     val expected_expression = ExpressionSchema(
       "test_expression",