+++ util-stats/src/test/scala/com/twitter/finagle/stats/RoleConfiguredStatsReceiverTest.scala
@@ -9,26 +9,26 @@ class RoleConfiguredStatsReceiverTest extends AnyFunSuite {
     val serverSR = new RoleConfiguredStatsReceiver(mem, Server)
 
     clientSR.counter("foo")
-    assert(mem.schemas.get(Seq("foo")).get.metricBuilder.role == Client)
+    assert(mem.schemas.get(Seq("foo")).get.role == Client)
     clientSR.addGauge("bar")(3)
-    assert(mem.schemas.get(Seq("bar")).get.metricBuilder.role == Client)
+    assert(mem.schemas.get(Seq("bar")).get.role == Client)
     clientSR.stat("baz")
-    assert(mem.schemas.get(Seq("baz")).get.metricBuilder.role == Client)
+    assert(mem.schemas.get(Seq("baz")).get.role == Client)
 
     serverSR.counter("fee")
-    assert(mem.schemas.get(Seq("fee")).get.metricBuilder.role == Server)
+    assert(mem.schemas.get(Seq("fee")).get.role == Server)
     serverSR.addGauge("fi")(3)
-    assert(mem.schemas.get(Seq("fi")).get.metricBuilder.role == Server)
+    assert(mem.schemas.get(Seq("fi")).get.role == Server)
     serverSR.stat("foe")
-    assert(mem.schemas.get(Seq("foe")).get.metricBuilder.role == Server)
+    assert(mem.schemas.get(Seq("foe")).get.role == Server)
   }
 
   test("RoleConfiguredSR equality") {
     val mem = new InMemoryStatsReceiver
-    val clientA = new RoleConfiguredStatsReceiver(mem, Client)
-    val clientB = new RoleConfiguredStatsReceiver(mem, Client)
-    val serverA = new RoleConfiguredStatsReceiver(mem, Server)
-    val serverB = new RoleConfiguredStatsReceiver(mem, Server)
+    val clientA = RoleConfiguredStatsReceiver(mem, Client)
+    val clientB = RoleConfiguredStatsReceiver(mem, Client)
+    val serverA = RoleConfiguredStatsReceiver(mem, Server)
+    val serverB = RoleConfiguredStatsReceiver(mem, Server)
 
     assert(clientA == clientB)
     assert(clientA != serverA)